pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-atomizer"
version = "0.1.0"
description = "ETL system for converting Word documents to structured knowledge atoms"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-docx>=0.8.11",
    "streamlit>=1.28.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "hypothesis>=6.92.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.hypothesis]
max_examples = 100
